Cleanup, make event_handler and packet_handler distinct concepts
