+++ login/src/packet_handler/delete_character.rs
